ia_service/requirements.txt
chromadb~=1.3.3
fastapi~=0.104
joblib~=1.5.2
langchain-community~=0.4.1
langchain-core~=1.0.3
langchain-google-genai~=3.0.0
langchain-ollama~=1.0.0
langchain-openai~=1.0.1
langgraph~=0.6
numpy~=2.2
ollama~=0.6.0
openai~=2.3
pandas~=2.3
pdfplumber~=0.11.7
pydantic~=2.11
pytest~=8.4
python-dotenv~=1.0
requests~=2.32
sentence-transformers~=5.1.2
# torch 2.5.1 is installed in the Dockerfile
transformers~=4.56.2
uvicorn[standard]~=0.24